feat: separate 'uninstall'
